+++ internal/config.go
@@ -3,6 +3,7 @@ package sf
 import (
 	"io/ioutil"
 	"os"
+	"strings"
 
 	"github.com/cloudflare/cloudflare-go"
 
@@ -77,18 +78,23 @@ func GetConfig(configFilePath string) Configuration {
 	// Set default values
 	if config.Cloudflare.Defaults.Type == "" {
 		config.Cloudflare.Defaults.Type = "CNAME"
+		log.Debug("Cloudflare default type is empty, defaulting to ", config.Cloudflare.Defaults.Type)
 	}
 	if config.Cloudflare.Defaults.TTL == 0 || config.Cloudflare.Defaults.Proxied {
 		config.Cloudflare.Defaults.TTL = 1
+		log.Debug("Cloudflare default TTL is empty, defaulting to ", config.Cloudflare.Defaults.TTL)
 	}
 	if config.IPProviders == nil {
 		config.IPProviders = append(config.IPProviders, "https://ifconfig.me/ip", "https://ipecho.net/plain", "https://myip.is/ip/")
+		log.Debug("IP providers is empty, defaulting to ", strings.Join(config.IPProviders, ", "))
 	}
 	if config.Notifications.Slack.Username == "" {
 		config.Notifications.Slack.Username = "SyncFlaer"
+		log.Debug("Slack username is empty, defaulting to ", config.Notifications.Slack.Username)
 	}
 	if config.Notifications.Slack.IconURL == "" {
 		config.Notifications.Slack.IconURL = "https://www.cloudflare.com/img/cf-facebook-card.png"
+		log.Debug("Slack icon URL is empty, defaulting to ", config.Notifications.Slack.IconURL)
 	}
 
 	// Validate config